move AccountWriter methods to HashingWriter and HistoryWriter
