+++ crates/stages/src/stages/index_account_history.rs
@@ -1,7 +1,7 @@
 use crate::{ExecInput, ExecOutput, Stage, StageError, UnwindInput, UnwindOutput};
 use reth_db::database::Database;
 use reth_primitives::stage::{StageCheckpoint, StageId};
-use reth_provider::{AccountExtReader, AccountWriter, DatabaseProviderRW};
+use reth_provider::{AccountExtReader, DatabaseProviderRW, HistoryWriter};
 use std::fmt::Debug;
 
 /// Stage is indexing history the account changesets generated in
