+++ crates/storage/provider/src/traits/account.rs
@@ -42,27 +42,3 @@ pub trait AccountExtReader: Send + Sync {
         range: RangeInclusive<BlockNumber>,
     ) -> Result<BTreeMap<Address, Vec<BlockNumber>>>;
 }
-
-/// Account reader
-#[auto_impl(&, Arc, Box)]
-pub trait AccountWriter: Send + Sync {
-    /// Unwind and clear account hashing
-    fn unwind_account_hashing(&self, range: RangeInclusive<BlockNumber>) -> Result<()>;
-
-    /// Unwind and clear account history indices.
-    ///
-    /// Returns number of changesets walked.
-    fn unwind_account_history_indices(&self, range: RangeInclusive<BlockNumber>) -> Result<usize>;
-
-    /// Insert account change index to database. Used inside AccountHistoryIndex stage
-    fn insert_account_history_index(
-        &self,
-        account_transitions: BTreeMap<Address, Vec<u64>>,
-    ) -> Result<()>;
-
-    /// iterate over accounts and insert them to hashing table
-    fn insert_account_for_hashing(
-        &self,
-        accounts: impl IntoIterator<Item = (Address, Option<Account>)>,
-    ) -> Result<()>;
-}
